requirements.txt
build==1.2.2.post1
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
contourpy==1.3.1
cycler==0.12.1
fonttools==4.55.5
idna==3.10
imageio==2.37.0
kiwisolver==1.4.8
lazy_loader==0.4
networkx==3.4.2
numpy==2.2.2
packaging==24.2
pillow==11.1.0
pip-autoremove==0.10.0
pip-tools==7.4.1
pyparsing==3.2.1
pyproject_hooks==1.2.0
PyQt5-Qt5==5.15.16
PyQt5_sip==12.16.1
python-dateutil==2.9.0.post0
requests==2.32.3
requirements-parser==0.11.0
scipy==1.15.1
setuptools==75.8.0
six==1.17.0
tifffile==2025.1.10
types-setuptools==75.8.0.20250110
urllib3==2.3.0
wheel==0.45.1
